Add emoji fetch restriction. Fix issue with missing messageId registration in database.

+++ bot.ts
@@ -1,26 +1,20 @@
-import { Client, GatewayIntentBits, Partials } from "discord.js";
+import { Client, GatewayIntentBits, MessageReaction, PartialMessageReaction, Partials } from "discord.js";
 import GiveawayService from "./lib/services/GiveawayService";
 import ParticipantService from "./lib/services/ParticipantService";
 import { getDiscordVariables } from "./lib/utils/server";
 
 const { token, giveawayChannelId, giveawayEmoji } = getDiscordVariables();
 
-const client = new Client({
-	intents: [ GatewayIntentBits.Guilds, GatewayIntentBits.GuildMembers, GatewayIntentBits.DirectMessages, GatewayIntentBits.GuildMessageReactions, GatewayIntentBits.DirectMessageReactions ],
-	partials: [ Partials.Message, Partials.Channel, Partials.Reaction, Partials.User ],
-});
-
-client.on("ready", async () => {
-	console.log("Discord Bot is ready");
-	await checkGiveawaysInterval(10);
-});
-client.login(token);
+const client = initializeClient();
 
 client.on("messageReactionAdd", async (reaction, user) => {
 	try {
-		if (reaction && reaction.message.id && reaction.message.channelId === giveawayChannelId && reaction.emoji.name === giveawayEmoji.toString()) {
+		const isBot = client.user?.id === user.id;
+		if (isBot) return;
+
+		if (reaction.message.channelId === giveawayChannelId && isRequiredEmoji(reaction))
 			await GiveawayService.addParticipant(reaction.message.id, user, reaction);
-		} else reaction.users.remove(user.id);
+		else reaction.users.remove(user.id);
 	} catch (error) {
 		console.error(error);
 		return null;
@@ -31,10 +25,8 @@ client.on("messageReactionRemove", async (reaction, user) => {
 		const giveaway = await GiveawayService.getByMessageId(reaction.message.id);
 		if (!giveaway) return;
 
-		if (reaction.message.id === giveaway.messageId) {
+		if (reaction.message.id === giveaway.messageId && isRequiredEmoji(reaction))
 			await ParticipantService.delete(giveaway.id, user.id);
-			await reaction.users.remove(user.id);
-		}
 	} catch (error) {
 		console.error(error);
 		return null;
@@ -50,5 +42,22 @@ async function checkGiveawaysInterval(intervalInMinutes: number = 5) {
 		await GiveawayService.checkGiveaways();
 		pending = false;
 	}, timeInMinutes);
+}
+function initializeClient() {
+	const client = new Client({
+		intents: [ GatewayIntentBits.Guilds, GatewayIntentBits.GuildMembers, GatewayIntentBits.DirectMessages, GatewayIntentBits.GuildMessageReactions, GatewayIntentBits.DirectMessageReactions ],
+		partials: [ Partials.Message, Partials.Channel, Partials.Reaction, Partials.User ],
+	});
+
+	client.on("ready", async () => {
+		console.log("Discord Bot is ready");
+		await checkGiveawaysInterval(10);
+	});
+
+	client.login(token);
+	return client;
+}
+function isRequiredEmoji(reaction: PartialMessageReaction | MessageReaction) {
+	return (reaction.emoji.id ?? reaction.emoji.name) === giveawayEmoji;
 }
 export default client;
